requirements.txt
anthropic>=0.42.0
PyGithub>=2.5.0
gitpython>=3.1.0
python-dotenv>=1.0.0
